Refactor socket server code and fix header format
This commit refactors the socket server code in the `serverImp.py` file. It removes the unused `csvobj` variable and fixes the format of the `headers` variable in the `ACK-EXCHANGE` section. The previous format was incorrect and has been updated to `HEADERS{}:TYPES{}`. This change improves the functionality and readability of the socket server.